Shuffle stuff around and clean some things up
